Update barotopic streamfunciton plots
This commit fixes the y axis, which was incorrectly offset by 320 km,
like the x axis.

It also makes the axes of each figure equal (appropriate for x-y plots).
The size of figures is changed accordingly.

It also creates versions of the Ocean1 and Ocean2 plots covering the
full domain as well as just the cavity.